+++ Assets/Plugins/UnityPurchasing/script/IAPDemo.cs
@@ -200,16 +200,22 @@ public class IAPDemo : MonoBehaviour, IStoreListener
 		builder.AddProduct("100.gold.coins", ProductType.Consumable, new IDs
 		{
 			{"100.gold.coins.mac", MacAppStore.Name},
+			{"000000596586", TizenStore.Name},
+
 		});
 
 		builder.AddProduct("500.gold.coins", ProductType.Consumable, new IDs
 		{
 			{"500.gold.coins.mac", MacAppStore.Name},
+			{"000000596581", TizenStore.Name},
+
 		});
 
 		builder.AddProduct("sword", ProductType.NonConsumable, new IDs
 		{
-			{"sword.mac", MacAppStore.Name}
+			{"sword.mac", MacAppStore.Name},
+			{"000000596583", TizenStore.Name},
+
 		});
 
 		builder.AddProduct("subscription", ProductType.Subscription, new IDs
@@ -229,6 +235,12 @@ public class IAPDemo : MonoBehaviour, IStoreListener
 		// A) Unity IAP does not automatically restore purchases on Samsung Galaxy Apps
 		// B) IAPDemo (this) displays the "Restore" GUI button for Samsung Galaxy Apps
 		m_IsSamsungAppsStoreSelected = module.androidStore == AndroidStore.SamsungApps;
+
+
+		// This selects the GroupId that was created in the Tizen Store for this set of products
+		// An empty or non-matching GroupId here will result in no products available for purchase
+		builder.Configure<ITizenStoreConfiguration>().SetGroupId("100000085616");
+
 
 		#if RECEIPT_VALIDATION
 		validator = new CrossPlatformValidator(GooglePlayTangle.Data(), AppleTangle.Data(), Application.bundleIdentifier);